index.html mainpage.css mainpage.js
refactored index.html to fix conflict with animations not working correctly and to correct issues for group merger.

restructured added secondary function to correct autoRun not working to allow page to operate on it's own.
** NOTE need to fix bug with backward arrow not working creating undefined condition.

**NOTE fix issue where autorun and arrows after cycling through does not reset.

changed several elements in .css that where defined incorrectly with incorrect class tree to fix conflicts with rendering on mainpage.